+++ src/responses/parameters.rs
@@ -61,7 +61,7 @@ pub struct GlobalRuntimeParameterValue(pub serde_json::Value);
 
 impl fmt::Display for GlobalRuntimeParameterValue {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        writeln!(f, "{}", &self.0)?;
+        writeln!(f, "{}", self.0)?;
 
         Ok(())
     }